+++ packages/e2e/lib/e2e/libe2e.e2e.dart
@@ -1,0 +1,35 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// dart format width=80
+
+// **************************************************************************
+// Generator: E2ELibrary
+// **************************************************************************
+
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint, type=warning
+import 'libe2e.dart' as _i1;
+import 'a.dart' as _i2;
+import 'b.dart' as _i3;
+import 'c.dart' as _i4;
+import 'd.dart' as _i5;
+import 'e.dart' as _i6;
+import 'f.dart' as _i7;
+import 'package:flutter/src/foundation/change_notifier.dart' as _i8;
+
+void fnE2e([
+  Object value = const _i1.Foo(),
+]) =>
+    throw UnimplementedError();
+void prefixToRelativePathE2e([
+  Object value = const _i2.A(),
+  Object value2 = const _i3.B(),
+  Object value3 = const _i4.C(),
+  Object value4 = const _i5.D(),
+  Object value5 = const _i6.E(),
+  Object value6 = const _i7.F(),
+]) =>
+    throw UnimplementedError();
+void notifierRegressionE2e([
+  _i8.ValueNotifier<int>? value,
+]) =>
+    throw UnimplementedError();